save dev state — impl get_latest_block
